playground/requirements.txt
httpx>=0.28.0
python-dotenv>=1.0.0
rich>=13.0.0
